confirm deleting alarms [ref #91723268]

+++ app/src/main/java/is/hello/sense/ui/fragments/SmartAlarmDetailFragment.java
@@ -34,6 +34,7 @@ import is.hello.sense.ui.dialogs.LoadingDialogFragment;
 import is.hello.sense.ui.dialogs.SmartAlarmSoundDialogFragment;
 import is.hello.sense.ui.dialogs.TimePickerDialogFragment;
 import is.hello.sense.ui.handholding.WelcomeDialogFragment;
+import is.hello.sense.ui.widget.SenseAlertDialog;
 import is.hello.sense.ui.widget.util.Views;
 import is.hello.sense.util.Analytics;
 import is.hello.sense.util.DateFormatter;
@@ -277,8 +278,16 @@ public class SmartAlarmDetailFragment extends InjectionFragment {
 
 
     public void deleteAlarm(@NonNull View sender) {
-        LoadingDialogFragment.show(getFragmentManager(), null, false);
-        bindAndSubscribe(smartAlarmPresenter.deleteSmartAlarm(index), ignored -> finish(), this::presentError);
+        SenseAlertDialog confirmDelete = new SenseAlertDialog(getActivity());
+        confirmDelete.setTitle(R.string.label_delete_alarm);
+        confirmDelete.setMessage(R.string.dialog_message_confirm_delete_alarm);
+        confirmDelete.setPositiveButton(R.string.action_delete, (dialog, which) -> {
+            LoadingDialogFragment.show(getFragmentManager(), null, false);
+            bindAndSubscribe(smartAlarmPresenter.deleteSmartAlarm(index), ignored -> finish(), this::presentError);
+        });
+        confirmDelete.setNegativeButton(android.R.string.cancel, null);
+        confirmDelete.setDestructive(true);
+        confirmDelete.show();
     }
 
     public void saveAlarm() {
